// src/app/models/instructor.ts
export class Instructor {
    id?: string;
    usuario?: string;
    nombre_usuario?: string;
    foto_usuario?: string;
    estado_foto?:string;
    perfil?: string;
    telefono?: string;
    fecha_create?: number;
    fecha_update?: number;
    usuario_update?: string;
    especialidades?:any[];
    firma?:string;
    estado_firma?:string;
    observaciones?:string;
    estado?:string;
    estado_solicitud?:string;
    max_especialidades?:number;
}